restructure project pages hierarchy
