+++ src/projects/project_pages/SuppTrack.vue
@@ -30,10 +30,10 @@
 </template>
 
 <script>
-import twoColumnTemplate from '../components/templates/twoColumnTemplate.vue'
-import SupptrackImage2 from '../projects/project_images/SupptrackImage2.png';
-import SupptrackImage3 from '../projects/project_images/SupptrackImage3.png';
-import SupptrackVideo from '../projects/project_videos/supptrackVideo.mp4';
+import twoColumnTemplate from '../../components/templates/twoColumnTemplate.vue'
+import SupptrackImage2 from '../project_images/SupptrackImage2.png';
+import SupptrackImage3 from '../project_images/SupptrackImage3.png';
+import SupptrackVideo from '../project_videos/supptrackVideo.mp4';
 
 export default {
     name: 'SuppTrack',
